Hack to avoid ClassNotFound errors in RC VMs.
With the upgrade to Selenium Grid 1.0.7 the RC VMs started
to throw this error:

21:57:26.499 INFO - Ping Hub at ...
Exception in thread "Thread-2" java.lang.NoClassDefFoundError:
...
Caused by: java.lang.ClassNotFoundException:
com.thoughtworks.selenium.grid.HttpClient

Selenium Grid now needs other classes that this plugin does
not package. The classes happen to be included in the hub jar,
so as a hack this change adds the hub jar to the classpath
of the RC VMs.